+++ .dtproject
@@ -1,0 +1,3 @@
+VERSION=0.1.0
+TYPE=template-basic
+TYPE_VERSION=1
