+++ duckietown_rl/env.py
@@ -2,6 +2,8 @@ import gym
 import gym_duckietown
 from gym_duckietown.envs.duckietown_env import DuckietownLF
 
+from wrappers import NormalizeWrapper, ImgWrapper, \
+    DtRewardWrapper, ActionWrapper, ResizeWrapper, SteeringToWheelVelWrapper
 
 def launch_env(id=None):
     env = None
@@ -20,5 +22,13 @@ def launch_env(id=None):
         )
     else:
         env = gym.make(id)
+
+    # Wrappers
+    env = ResizeWrapper(env)
+    env = NormalizeWrapper(env)
+    env = ImgWrapper(env)  # to make the images from 160x120x3 into 3x160x120
+    env = SteeringToWheelVelWrapper(env)
+    env = ActionWrapper(env)
+    env = DtRewardWrapper(env)
 
     return env
