+++ solution.py
@@ -4,43 +4,10 @@ from typing import Tuple
 
 import numpy as np
 
-from aido_schemas import EpisodeStart, protocol_agent_duckiebot1, PWMCommands, Duckiebot1Commands, LEDSCommands, RGB, \
-    wrap_direct, Context, Duckiebot1Observations, JPGImage
+from aido_schemas import (Context, Duckiebot1Commands, Duckiebot1Observations, EpisodeStart, LEDSCommands, protocol_agent_duckiebot1, PWMCommands, RGB, wrap_direct)
 
 from model import DDPG
 from wrappers import *
-
-class FakeWrap:
-    def __init__(self):
-        self.env = None
-        self.action_space = None
-
-        self.camera_width = 640
-        self.camera_height = 480
-        self.observation_space = spaces.Box(
-                low=0,
-                high=255,
-                shape=(self.camera_height, self.camera_width, 3),
-                dtype=np.uint8
-        )
-        self.reward_range = None
-        self.metadata = None
-
-
-class DTPytorchWrapper():
-    def __init__(self, shape=(120, 160, 3)):
-        self.shape = shape
-        self.transposed_shape = (shape[2], shape[0], shape[1])
-
-    def preprocess(self, obs):
-        # from PIL import Image
-        # return np.array(Image.fromarray(obs).resize(self.shape[0:2])).transpose(2, 0, 1)
-
-        import cv2
-        obs = cv2.resize(obs, self.shape[0:2])
-        # NOTICE: OpenCV changes the order of the channels !!!
-        obs = cv2.cvtColor(obs, cv2.COLOR_BGR2RGB)
-        return obs.transpose(2, 0, 1)
 
 class PytorchRLBaseline:
     def __init__(self, load_model=False, model_path=None):
@@ -50,10 +17,7 @@ class PytorchRLBaseline:
         self.model = DDPG(state_dim=self.image_processor.shape, action_dim=2, max_action=1, net_type="cnn")
         self.current_image = np.zeros((640, 480, 3))
 
-        # TODO: Uncomment when you train your model!
-
-        # fp = model_path if model_path else "model"
-        # self.model.load(fp, "models", for_inference=True)
+        self.model.load("model", directory="./models")
 
     def init(self, context: Context):
         context.info('init()')
@@ -65,7 +29,7 @@ class PytorchRLBaseline:
         context.info(f'Starting episode "{data.episode_name}".')
 
     def on_received_observations(self, data: Duckiebot1Observations):
-        camera: JPGImage = data.camera
+        camera = data.camera
         obs = jpg2rgb(camera.jpg_data)
         self.current_image = self.image_processor.preprocess(obs)
